Refactor(gallery): Re-org button and dialog
Nothing much, just renaming and moving around to make more sense
